refactor(reconciliation): split god-module into config/scoring/stats/matching
reconciliation.py (1438 lines) mixed four concerns. Split into focused modules:
- reconciliation_config.py  <- ReconciliationConfig, MatchCandidate, entry helpers, load_reconciliation_config
- reconciliation_scoring.py <- score_* / normalize_text / weighted_total / ai_semantic_score
- reconciliation_stats.py   <- ReconciliationStats + get_reconciliation_stats
- reconciliation.py keeps the matching engine + re-exports the moved names so
  `from src.services.reconciliation import …` call sites are unchanged.

The config/scoring/matching layering avoids a circular import (scoring needs
ReconciliationConfig; matching needs both). Two tests repointed their
monkeypatch targets to where the moved functions now look up settings/Path.

reconciliation.py: 1438 -> 917. Structural (total LOC ~flat). 219 reconciliation
tests green; ruff clean.
